requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-mssql-backend==2.8.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==3.4.0
importlib-resources==5.1.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.6.0
pylint==2.6.0
pymssql==2.2.1
pyodbc==4.0.30
pytz==2020.5
six==1.15.0
sqlparse=0.4.2
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.4.2
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
